finising uno
wild cards were made black, everything is now in relation to canvas.width/height and the text color on the cards was changed so that it is visible on black. also the button that you click to take a card and end your turn disipears after you play a wild.